Fixed bug with considerLambdas
Newly implemented considerLambdas was activating any state that had a lambda transition into it, regardless of if the previous state was live or not. Fixed bug by editing conditional statement.

Also added lambdaStates field to NFA class.
The lambdaStates class has a length equal to the amount of states which have a lambda transition and its elements are the indexes of such states. This is done so that whenever we update the states based on lambda transitions, instead of looking at n^2 states where n is all states in the NFA, we are only looking at where n is the amount of states with only a lambda transition, which likely results in a much smaller n^2 value.