Fixed logic issues with digits and underscores
Encountered some errors where the input of an underscore and only a SINGLE integer would kill off Q2 and keep Q0 alive, when both states should have been alive for processing logic when a valid (underscore + integer) was inputted. All underscores in a Python Floating Point Literal must be resolved by 1 or more digits between 0-9 following it, but our current iteration of the NFA wouldn't process Q2's logic if the underscore was resolved by just a single integer, as there was only a transition between Q11 and Q0 when a valid digit was inputted.
*Resolved the issue relatively easily, by making a transition between Q11 and Q2 such that a valid underscore resolution via at least 1 digit of 0-9 also activated Q2 in addition to Q0.