Finished Part 1 and made Combined NFA
Completed Part 1 NFA and Java Representation

Also made a gigantic NFA which combined all of the other NFA's constructed to read in Decimal, Octal, Hexadecimal, and Floating Point values. Plugged in this new NFA into the Java code, and ran tests with no test runs resulting in a "failed" output.